Rename the SnapshotAnalyser to be more generic
