+++ src/Analyser.js
